fixed: - dont need []cityList - readme.md - promblem with same locations for refresh added: -method POST for add new locations -.env configuration file -forecasts automatically refresh -paralleled the process of getting weather data from an external API
