+++ weather/README.md
@@ -28,16 +28,14 @@
 
 
 -----------------------------------------------------------------------
-
-
-Информация о названиях городов и их координаты записывается в базу данных автоматически при запуске сервиса (в main.go файле находится
-[]cityList).
+Реализация:
+Сервис предназначен для получения, хранения и предоставления инфомации о прогнозе погоды (подробнее в EndPoints list) на предстоящие 5 дней в городах список которых используется [geocoding-api].
+При передаче в метод POST названия города отправляется запрос на сторонее API, из ответа на запрос извлекается вся небходимая информация в том числе координаты и формируется запрос в [open weather map] API откуда извлекается нужная информация о погоде. После чего данные обновляется каждые 30 секунд.
+Запись дынных ведется в Postgresql (locations, forecast). 
 
 В файле .env содержится необходимая конфигурация для запуска сервиса (может читаться из переменных окружения), в том числе и API key  для работы со сторонним сервисом
 
-
-
-EndPoints list for User:
+EndPoints list:
 
 (GET) "/weather/locations" -  list of cities, which we can get forcast
